+++ src/modules/Preservation/views/ScopeOverview/DialogTable.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { PreservedTag } from './types';
 import { Container } from './DialogTable.style';
-import ProcosysTable from '@procosys/components/Table/ProcosysTable';
+import ProcosysTable from '@procosys/components/Table';
 
 interface TableProps {
     tags: PreservedTag[];
